// src/components/jtl/jtl-dropdown/types/DropdownItem.ts
/**
 * Enum for different dropdown item types
 */
enum DropdownItem {
  Default = 'default',
  // ToggleGroup = 'toggleGroup', // TODO: wait for Toggle Component to be ready
  SubMenuTrigger = 'subMenuTrigger',
  Separator = 'separator',
  Label = 'label',
  Checkbox = 'checkbox',
}

export default DropdownItem;